+++ src/i18n.ts
@@ -1,41 +1,37 @@
-(() => {
-  window.raisinAddon = {
-    i18n: {
-      nl: {
+const i18n: any = {
+    nl: {
         giveNameChip: "Geef een naam",
         giveNameLabel: "Geef een naam op voor de rekening",
         giveNameLabelLengthError: "Naam moet meer dan 3 tekens zijn!",
         interestAccruedThisQuarter: "Opgebouwd dit kwartaal",
         totalInterestPaidOut: "Totaal uitbetaalde rente",
-      },
-      de: {
+    },
+    de: {
         giveNameChip: "Geben Sie einen Namen ein",
         giveNameLabel: "Geben Sie einen Namen für das Konto ein",
         giveNameLabelLengthError: "Name sollte mindestens 3 Zeichen sein!",
         interestAccruedThisQuarter: "Diesem Viertel gebaut",
         totalInterestPaidOut: "Gesamtzinsen bezahlt",
-      },
-      en: {
+    },
+    en: {
         giveNameChip: "Enter a name",
         giveNameLabel: "Enter a name for the account",
         giveNameLabelLengthError: "Name should be at least 3 characters!",
         interestAccruedThisQuarter: "Accrued this quarter",
         totalInterestPaidOut: "Total interest paid",
-      },
-      fr: {
+    },
+    fr: {
         giveNameChip: "Entrez un nom",
         giveNameLabel: "Entrez un nom pour le compte",
         giveNameLabelLengthError: "Le nom doit comporter au moins 3 caractères!",
         interestAccruedThisQuarter: "Construit ce trimestre",
         totalInterestPaidOut: "Intérêt total payé",
-      },
-      es: {
+    },
+    es: {
         giveNameChip: "Ingrese un nombre",
         giveNameLabel: "Ingrese un nombre para la cuenta",
         giveNameLabelLengthError: "¡El nombre debe ser al menos 3 caracteres!",
         interestAccruedThisQuarter: "Acumulado este trimestre",
         totalInterestPaidOut: "Intereses totales pagados",
-      },
     },
-  };
-})();
+}
