+++ tsconfig.json
@@ -3,6 +3,7 @@
        "strict": true,
        "module": "commonjs",
        "target": "es6",
+       "lib": ["ES2022", "DOM"],
        "esModuleInterop": true,
        "sourceMap": true,
        "rootDir": "src",
